requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
jsonschema==3.2.0
Markdown==3.3.3
pyrsistent==0.17.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
